Added tests for newly added interceptor
